fix: resolve gosec findings (G115 overflow + G204/G702/G703 false positives)
- disks.go: use float64 arithmetic to avoid uint64->int overflow in pct calc (G115)
- config.go: #nosec G703 -- path is user-supplied talosconfig flag by design
- machineconfig.go: #nosec G204 G702 -- editor comes from $EDITOR env var by design
- client.go: #nosec G204 on 4 talosctl invocations -- binary name is hardcoded

+++ internal/ui/disks.go
@@ -153,7 +153,7 @@ func renderVolumeRow(v talos.VolumeInfo, width int) string {
 	var detail string
 	if v.Available > 0 {
 		used := v.Size - v.Available
-		pct := int(used * 100 / v.Size)
+		pct := int(float64(used) * 100 / float64(v.Size))
 		bar := usageBar(pct, 12)
 		detail = fmt.Sprintf("%s  %s / %s (%d%%)",
 			bar,
